requirements.txt
django
six
pillow
crispy-bootstrap5
django_crispy_forms